requirements.txt
fastembed==0.7.3
langchain==1.0.0a9
langchain_cohere==0.4.6
langchain_openai==0.3.33 
langchain_qdrant==0.2.1
pymongo==4.15.1
python-dotenv==1.1.1
tavily-python==0.7.12
fastapi==0.118.0
uvicorn==0.37.0